Lawrence/qwen2.5 vl/encoder tests
* Add test file for vision encoder

* fix: reshape error in vision encoder rope

* rope in Qwen does not apply to adjancent dimensions but instead
  mirrored dimensions
* the head dimension is split in half
* for example, if the head dimension is 80, then the rope pairs are
  (0,40), (1,41), ...
* added an abundnace of test cases and tensor saves

* cleanup

---------

+++ torchtune/models/qwen2_5_vision/_transform.py
@@ -95,7 +95,7 @@ class Qwen2_5_VLImageTransform:
         size: Optional[Dict[str, int]] = None,
         min_pixels: Optional[int] = None,
         max_pixels: Optional[int] = None,
-        dtype: torch.dtype = torch.bfloat16,
+        dtype: torch.dtype = torch.float32,
         resample: str = "bicubic",
     ) -> None:
         self.patch_size = patch_size
@@ -262,7 +262,7 @@ class Qwen2_5_VLTransform(ModelTokenizer, Transform):
         max_seq_len: Optional[int] = None,
         image_mean: Optional[List[float]] = None,
         image_std: Optional[List[float]] = None,
-        dtype: torch.dtype = torch.bfloat16,
+        dtype: torch.dtype = torch.float32,
         prompt_template: Optional[_TemplateType] = None,
     ):
         special_tokens = (
